resolve link in self path

+++ deploy.sh
@@ -11,10 +11,16 @@ fn_stoponerror ()
 	fi
 }
 
-path="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+path=$(dirname "$(realpath "$0")")
+fn_stoponerror $BASH_SOURCE $LINENO $?
+echo path is $path
+if [ -z $path ]
+then
+	echo "path is empty"
+	exit
+fi
 
 path_p=$path'/httpd_portable'
-
 sudo pkill httpd_portable
 
 if [ -e $path_p ]
